Add bitsandbytes installation instructions and update dependencies
The `bitsandbytes` package is installed separately, since it is not
available on all devices. This package is needed for the train and
predict actions for some of the models. So the preprocessing and
evaluation (if you have the predictions already) tasks for all models
can be run without the `bitsandbytes` package.